requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.7.0
Flask-WTF==0.14.3
Pillow==8.1.0
email_validator==1.1.2
gunicorn==20.0.4
gevent==21.1.2
requests==2.25.1
pyOpenSSL==20.0.1
oauthlib==3.1.0
python-dotenv
